Convert test output listener test to new junit multiversion test
Split off TestNG test into a TestNG multiversion test
